+++ src/test/kotlin/tornadofx/testapps/DataGridTest.kt
@@ -1,5 +1,6 @@
 package tornadofx.testapps
 
+import javafx.collections.FXCollections
 import javafx.scene.control.SelectionMode
 import tornadofx.*
 import tornadofx.testapps.DataGridTestApp.Companion.images
@@ -30,6 +31,7 @@ class DataGridTestApp : App(DataGridTest::class, DataGridStyles::class) {
 
 class DataGridTest : View("DataGrid") {
     var datagrid: DataGrid<String> by singleAssign()
+    val list = FXCollections.observableArrayList<String>()
 
     override val root = borderpane {
         left {
@@ -37,21 +39,21 @@ class DataGridTest : View("DataGrid") {
                 combobox(values = images.keys.toList()) {
                     promptText = "Select images"
                     valueProperty().onChange {
-                        datagrid.items.setAll(images[it])
+                        list.setAll(images[it])
                     }
                     shortcut("k") { value = "kittens" }
                     shortcut("p") { value = "puppies" }
                 }
                 button("Add").action {
-                    datagrid.items.add("http://i.imgur.com/bvqTBT0b.jpg")
+                    list.add("http://i.imgur.com/bvqTBT0b.jpg")
                 }
             }
         }
         center {
-            datagrid = datagrid<String> {
+            datagrid = datagrid(list) {
                 setPrefSize(550.0, 550.0)
 
-                selectionModel.selectionMode = SelectionMode.MULTIPLE
+                selectionModel.selectionMode = SelectionMode.SINGLE
                 //maxCellsInRow = 3
                 //maxRows = 3
                 cellWidth = 164.0
@@ -68,6 +70,9 @@ class DataGridTest : View("DataGrid") {
         }
         bottom {
             label(stringBinding(datagrid.selectionModel.selectedItems) { joinToString(", ") })
+            button("Remove from index 2").action {
+                list.removeAt(2)
+            }
         }
     }